requirements.txt
# Legislative Tracker Dependencies

# Core
dash==3.1.1
dash-bootstrap-components==2.0.4
dash-ag-grid==33.3.3
plotly==6.2.0

# Data
pandas==2.3.1
pyarrow==21.0.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.9.0
requests>=2.32.0
PyYAML>=6.0

# Azure
adlfs>=2024.2.0
azure-identity==1.16.0
azure-storage-blob==12.20.0
azure-keyvault-secrets==4.7.0

# Production server
gunicorn==21.2.0
